+++ src/memory/file_loader.h
@@ -5,6 +5,6 @@
 #include <string.h>
 #include <stdio.h>
 
-int load_file(char* file, uint16_t file_offset, uint8_t* dest, uint16_t dest_offset, uint16_t max);
+int load_file(char* file, long file_offset, uint8_t* dest, long dest_offset, long max);
 
 #endif
